+++ package.js
@@ -1,7 +1,7 @@
 Package.describe({
   name: "ma:wizard",
   summary: "Forms and wizards management made easy",
-  version: "2.3.0",
+  version: "2.4.0",
   git: "https://github.com/doubleslashG/ma-wizard.git"
 });